feat:modalConfigs.js로 컴포넌트 분리
